requirements.txt
bcrypt==4.1.2
attrs==22.1.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
flask-mailman==1.1.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask-Bcrypt==1.0.1
fonttools==4.47.2
iniconfig==1.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==21.3
pandas==2.2.0
pillow==10.2.0
pluggy==1.0.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
six==1.16.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2023.4
Werkzeug==3.0.0
